Move scanner helper methods to it's own class
